requirements.txt
# External dependencies
tqdm
pyhmmer
biopython
pandas
muscle
numpy